Added user_exercise_history table and initialized DB schema
